Écriture du premier draft complété
L'objectif est de tenir la main pour certain concept, comme la création
de la première classe, les attributs, les constructeurs, et les méthodes
(que l'on voit avec la première classe `Tache`).

Par la suite, on attaque le destructeur dans la classe `Personne`
considérant qu'elle possède un pointeur (qui sera utilisé comme tableau
dynamique réalouable).

Enfin on laisse l'étudiant cogité un peu plus en modifiant le
laboratoire 1 en utilisant les nouveaux prototypes utilisant également
une classe `Menu` incluant un tableau statique.

Certaines des méthodes que nous avons fait deviendrons des *getter* et
des *setter*.
