+++ laboratoires/lab02-oop1/solution/vieux/fonctions.h
