+++ laboratoires/lab02-oop1/solution/vieux/menu.h
@@ -18,8 +18,6 @@ private:
 
     int selection;
 
-    bool aggrandirListeOptions(void);
-
 public:
 
     Menu(const std::string &, const std::string &);
